experiments/requirements.txt
Cython
jupyter
numpy
pandas<=1.2.3

#-f https://download.pytorch.org/whl/torch_stable.html
#torch==1.8.1+cpu
torchvision==0.9.1

PennyLane

qulacs-gpu
PennyLane-qulacs

PennyLane-qiskit==0.11.0
qiskit==0.23.6
qiskit-aer==0.7.5
qiskit-aqua==0.8.2
qiskit-ibmq-provider==0.11.1
qiskit-ignis==0.5.2
qiskit-terra==0.16.4
scikit-learn<=0.24.1
scipy<=1.6.1
seaborn==0.10.1
sympy==1.7.1
pyeda
tensorboard

pathos
pyyaml
tqdm
